+++ CelleGames.ProductApi/Repository/Interface/ICategoryRepository.cs
@@ -5,9 +5,9 @@ namespace CelleGames.ProductApi.Repository.Interface
     public interface ICategoryRepository
     {
         Task<IEnumerable<CategoryVO>> FindAll();
-        Task<CategoryVO> FindById(long id);
+        Task<CategoryVO> FindById(int id);
         Task<CategoryVO> Create(CategoryVO categoryVO);
         Task<CategoryVO> Update(CategoryVO categoryVO);
-        Task<bool> Delete(long id);
+        Task<bool> Delete(int id);
     }
 }
